+++ tests/rspack-plugin/helpers/runRspackWithPlugin.js
@@ -11,9 +11,11 @@
  *     outputPath: string,
  *     isServer: boolean,
  *     clientManifestFilename?: string,
+ *     clientReferenceDiagnosticsFilename?: string|false,
  *     clientReferences?: unknown,
  *     publicPath?: string,
  *     crossOriginLoading?: false|'anonymous'|'use-credentials',
+ *     withCss?: boolean,
  *     configExtra?: object,
  *   }
  *
@@ -41,9 +43,11 @@ const {
   outputPath,
   isServer,
   clientManifestFilename,
+  clientReferenceDiagnosticsFilename,
   clientReferences: rawClientReferences,
   publicPath,
   crossOriginLoading,
+  withCss,
   configExtra,
 } = args;
 
@@ -69,6 +73,23 @@ if (missingRuntimeEntries.length > 0) {
 }
 const runtimeEntry = isServer ? runtimeEntries.server : runtimeEntries.client;
 const clientReferences = reviveFromRunner(rawClientReferences);
+const revivedConfigExtra = reviveFromRunner(configExtra);
+const plugins = [
+  new RSCRspackPlugin({
+    isServer: isServer,
+    clientManifestFilename: clientManifestFilename,
+    clientReferenceDiagnosticsFilename: clientReferenceDiagnosticsFilename,
+    clientReferences: clientReferences,
+  }),
+];
+if (withCss) {
+  plugins.push(
+    new rspack.CssExtractRspackPlugin({
+      filename: '[name].css',
+      chunkFilename: '[name].chunk.css',
+    }),
+  );
+}
 
 const config = {
   mode: 'development',
@@ -88,14 +109,15 @@ const config = {
     minimize: false,
   },
   devtool: false,
-  plugins: [
-    new RSCRspackPlugin({
-      isServer: isServer,
-      clientManifestFilename: clientManifestFilename,
-      clientReferences: clientReferences,
-    }),
-  ],
-  ...(configExtra || {}),
+  ...(withCss
+    ? {
+        module: {
+          rules: [{ test: /\.css$/, use: [rspack.CssExtractRspackPlugin.loader, 'css-loader'] }],
+        },
+      }
+    : {}),
+  plugins,
+  ...(revivedConfigExtra || {}),
 };
 
 rspack(config, (err, stats) => {