+++ tests/webpack-plugin/fixtures/static-islands/TinyIsland.js
@@ -1,0 +1,5 @@
+'use client';
+
+export default function TinyIsland() {
+  return 'tiny interactive island';
+}